+++ src/main/java/sum25/studentcode/backend/modules/Level/dto/response/LevelResponse.java
@@ -12,7 +12,7 @@ public class LevelResponse {
 
     private Long levelId;
     private String levelName;
-    private Integer difficultyScore;
+    private Double difficultyScore;
     private String description;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
